feat: 添加api

+++ src/service/controller/service.ts
@@ -1,7 +1,8 @@
 import { db } from './db'
 import { TableName, SearchParameters } from '../model'
 import dayjs from 'dayjs'
-
+import { useOnline } from '@vueuse/core'
+const isOnline = useOnline()
 export async function saveStore<
   T extends {
     createTime?: string
@@ -9,26 +10,30 @@ export async function saveStore<
   }
 >(tableName: TableName, datas: T, id?: number) {
   const now = dayjs().format('YYYY-MM-DD HH:mm:ss')
-  if (id) {
-    const ret = await db.table(tableName).update(id, { ...datas, updateTime: now })
-    return ret
-  } else {
-    const ret = await db.table(tableName).add({ ...datas, createTime: now })
-    return ret
-  }
+  if (!isOnline.value) return 
+    if (id) {
+      const ret = await db.table(tableName).update(id, { ...datas, updateTime: now })
+      return ret
+    } else {
+      const ret = await db.table(tableName).add({ ...datas, createTime: now })
+      return ret
+    }
 }
 
 export async function deleteStore(tableName: TableName, ids: number[]) {
+  if (!isOnline.value) return 
   const ret = await db.table(tableName).bulkDelete(ids)
   return ret
 }
 
 export async function getStore<T>(tableName: TableName, key: number) {
+  if (!isOnline.value) return 
   const ret = await db.table(tableName).get(key)
   return ret as T
 }
 
 export async function getAllStore<T>(tableName: TableName) {
+  if (!isOnline.value) return 
   const ret = await db.table(tableName).toArray()
   return ret as T[]
 }
@@ -39,6 +44,7 @@ export async function filterStore<T>(
   pageNo = 0,
   pageSize = 10
 ) {
+  if (!isOnline.value) return 
   const ret = await db
     .table(tableName)
     .orderBy('id')
@@ -49,6 +55,7 @@ export async function filterStore<T>(
   return ret as T[]
 }
 export async function searchStore<T>(tableName: TableName, p?: SearchParameters<T>, all = false) {
+  if (!isOnline.value) return 
   const fuzzyQuery = (d: any, KV: [string, any][]) => {
     for (const [key, value] of KV) {
       const reg = new RegExp(value)
@@ -90,6 +97,7 @@ export async function searchStore<T>(tableName: TableName, p?: SearchParameters<
 }
 /**获取字符数量 */
 export async function getCount(tableName: TableName) {
+  if (!isOnline.value) return 
   const ret = await db.table(tableName).get('count')
   const ret2 = await db.table(tableName).get('updateTime')
   return {
